requirements.txt
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
lxml==5.0.0
imantics
segment_anything
onnx
onnxruntime
opencv-python
pycocotools
matplotlib
pyyaml
torch
torchvision
torchaudio
tqdm